[fix]: fix incorrect directory path
개발 중 잘못 이동된 파일 이동
