+++ src/main/java/com/pinkdumbell/cocobob/domain/auth/UserDetailService.java
@@ -1,7 +1,7 @@
 package com.pinkdumbell.cocobob.domain.auth;
 
-import com.pinkdumbell.cocobob.domain.product.user.User;
-import com.pinkdumbell.cocobob.domain.product.user.UserRepository;
+import com.pinkdumbell.cocobob.domain.user.User;
+import com.pinkdumbell.cocobob.domain.user.UserRepository;
 import com.pinkdumbell.cocobob.exception.CustomException;
 import com.pinkdumbell.cocobob.exception.ErrorCode;
 import java.util.ArrayList;
