+++ src/main/java/com/pinkdumbell/cocobob/domain/pet/PetService.java
@@ -7,9 +7,9 @@ import com.pinkdumbell.cocobob.domain.pet.dto.PetCreateRequestDto;
 import com.pinkdumbell.cocobob.domain.pet.dto.PetCreateResponseDto;
 import com.pinkdumbell.cocobob.domain.pet.image.PetImage;
 import com.pinkdumbell.cocobob.domain.pet.image.PetImageRepository;
-import com.pinkdumbell.cocobob.domain.product.user.User;
-import com.pinkdumbell.cocobob.domain.product.user.UserRepository;
-import com.pinkdumbell.cocobob.domain.product.user.dto.LoginUserInfo;
+import com.pinkdumbell.cocobob.domain.user.User;
+import com.pinkdumbell.cocobob.domain.user.UserRepository;
+import com.pinkdumbell.cocobob.domain.user.dto.LoginUserInfo;
 import com.pinkdumbell.cocobob.exception.CustomException;
 import com.pinkdumbell.cocobob.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
