+++ src/main/java/com/pinkdumbell/cocobob/domain/product/dto/PetPropertyResponseDto.java
@@ -1,5 +1,7 @@
 package com.pinkdumbell.cocobob.domain.product.dto;
 
+import com.pinkdumbell.cocobob.domain.product.property.PetProperty;
+import com.pinkdumbell.cocobob.domain.product.property.PetPropertyRepository;
 import com.pinkdumbell.cocobob.domain.product.property.Property;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,4 +11,10 @@ import lombok.Getter;
 public class PetPropertyResponseDto {
 
     private Property property;
+
+    public PetPropertyResponseDto(PetProperty property) {
+        this.property = property.getProperty();
+    }
+
+
 }
