+++ src/main/java/com/pinkdumbell/cocobob/domain/product/like/Like.java
@@ -1,7 +1,7 @@
 package com.pinkdumbell.cocobob.domain.product.like;
 
 import com.pinkdumbell.cocobob.domain.product.Product;
-import com.pinkdumbell.cocobob.domain.product.user.User;
+import com.pinkdumbell.cocobob.domain.user.User;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
