+++ src/main/java/com/pinkdumbell/cocobob/domain/user/dto/UserLoginRequestDto.java
@@ -1,15 +1,17 @@
-package com.pinkdumbell.cocobob.domain.product.user.dto;
+package com.pinkdumbell.cocobob.domain.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class UserEmailRequestDto {
-    String email;
+public class UserLoginRequestDto {
+
+    private String email;
+    private String password;
+
 }
