+++ src/test/java/com/pinkdumbell/cocobob/domain/user/UserRepositoryTest.java
@@ -1,8 +1,6 @@
 package com.pinkdumbell.cocobob.domain.user;
 
 import com.pinkdumbell.cocobob.config.JpaAuditingConfig;
-import com.pinkdumbell.cocobob.domain.product.user.User;
-import com.pinkdumbell.cocobob.domain.product.user.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
