+++ src/test/java/com/pinkdumbell/cocobob/domain/user/UserServiceTest.java
@@ -4,12 +4,9 @@ import com.pinkdumbell.cocobob.common.EmailUtil;
 import com.pinkdumbell.cocobob.common.dto.EmailSendResultDto;
 import com.pinkdumbell.cocobob.config.MailConfig;
 import com.pinkdumbell.cocobob.domain.auth.JwtTokenProvider;
-import com.pinkdumbell.cocobob.domain.product.user.User;
-import com.pinkdumbell.cocobob.domain.product.user.UserRepository;
-import com.pinkdumbell.cocobob.domain.product.user.UserService;
-import com.pinkdumbell.cocobob.domain.product.user.dto.UserCreateRequestDto;
-import com.pinkdumbell.cocobob.domain.product.user.dto.UserCreateResponseDto;
-import com.pinkdumbell.cocobob.domain.product.user.dto.UserEmailRequestDto;
+import com.pinkdumbell.cocobob.domain.user.dto.UserCreateRequestDto;
+import com.pinkdumbell.cocobob.domain.user.dto.UserCreateResponseDto;
+import com.pinkdumbell.cocobob.domain.user.dto.UserEmailRequestDto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
